refactor: signup modal to tsx
